requirements.txt
awscli==1.16.273
boto3==1.10.9
botocore==1.13.9
colorama==0.3.3
docutils==0.12
ecdsa==0.13.3
Fabric3==1.14.post1
futures==3.0.5
jmespath==0.9.0
paramiko==2.0.9
pyasn1==0.1.9
python-dateutil==2.6.1
pyyaml==4.2b1
rsa==3.3
s3transfer==0.2.1
six==1.11.0
tabulate==0.7.5
